refact types.Diagnostic to diagnostics.py
